+++ inst/include/ensmallen_bits/fbs/fbs_impl.hpp
@@ -77,8 +77,11 @@ FBS<BackwardStepType>::Optimize(FunctionType& function,
   // Controls early termination of the optimization process.
   bool terminate = false;
 
+  const size_t actualMaxIterations = (maxIterations == 0) ?
+      std::numeric_limits<size_t>::max() : maxIterations;
+
   Callback::BeginOptimization(*this, f, iterate, callbacks...);
-  for (size_t i = 1; i != maxIterations && !terminate; ++i)
+  for (size_t i = 0; i < actualMaxIterations && !terminate; ++i)
   {
     // During this optimization, we want to optimize h(x) = f(x) + g(x).
     // f(x) is `f`, but g(x) is specified by `BackwardStepType`.