+++ NEWS.md
@@ -1,3 +1,14 @@
+# RcppEnsmallen 0.3.11.0.1
+
+- Upgraded to ensmallen 3.11.0: "Sunny Day" (2025-12-16)
+  - Refactor `GradientDescent` into
+    `GradientDescentType<UpdatePolicyType, DecayPolicyType>` and
+    add the `DeltaBarDelta` and `MomentumDeltaBarDelta` optimizers
+    ([#440](https://github.com/mlpack/ensmallen/pull/440)).
+  - Fix an off-by-one bug where the actual number of executed iterations was one
+    fewer than the specified `maxIterations`
+    ([#443](https://github.com/mlpack/ensmallen/pull/443)).
+
 # RcppEnsmallen 0.3.10.0.1
 
 - Upgraded to ensmallen 3.10.0: "Unexpected Rain" (2025-09-30)